Split sensor drivers into separate files
